removed white space and unused variables
