+++ client/src/components/Footer.jsx
@@ -2,28 +2,23 @@ import React from 'react';
 import '../styles/footer.css';
 
 export const Footer = () => {
-    return (
-        
-  <div className='wrapper'>
-
-      
-      <div className='col-1'>
+  return (
+    <div className="wrapper">
+      <div className="col-1">
         <h3>Community</h3>
         <ul className="list-layout">
           <li className="link-contrast">About</li>
           <li className="link-contrast">Photo Gallery</li>
         </ul>
       </div>
-
-      <div className='col-1'>
+      <div className="col-1">
         <h3>Safety</h3>
         <ul className="list-layout">
           <li className="link-contrast">Trust &amp; Basics</li>
           <li className="link-contrast">Tips</li>
         </ul>
       </div>
-
-      <div className='col-1'>
+      <div className="col-1">
         <h3>Help and Support</h3>
         <ul className="list-layout">
           <li className="link-contrast">Contact Us</li>
@@ -32,7 +27,7 @@ export const Footer = () => {
         </ul>
       </div>
     </div>
-    )
-}
+  );
+};
 
 export default Footer;
