+++ client/src/components/Listings.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { useContext } from 'react';
 import { AppContext } from '../context/AppContext';
 import Search from './Search';
 import { useHistory } from 'react-router-dom';
