+++ client/src/components/Navigation.jsx
@@ -3,7 +3,7 @@ import '../styles/nav.css';
 import { Link } from 'react-router-dom';
 import { AppContext } from '../context/AppContext';
 import Logout from './Logout';
-import logo from '../images/housed.png'
+import logo from '../images/housed.png';
 
 export const Navigation = () => {
   const { user, loggedIn, setLoggedIn } = useContext(AppContext);
@@ -18,13 +18,8 @@ export const Navigation = () => {
   console.log(loggedIn);
   console.log(user);
 
-
-
   return (
     <div className="navbar">
-
-
-
       <Link to="/">
         <div className="logo">
           <img src={logo} alt="Housed logo" />
@@ -32,12 +27,7 @@ export const Navigation = () => {
       </Link>
       <div className="nav-right">
         {loggedIn ? (
-          
-            <Logout />
-          
-           
-          
-   
+          <Logout />
         ) : (
           <>
             <Link to="/howitworks"> How it works</Link>
