+++ package.json
@@ -18,6 +18,7 @@
     "moment-timezone": "^0.5.31",
     "mongoose": "^5.9.16",
     "multer": "^1.4.2",
+    "prettier": "^2.0.5",
     "react": "^16.13.1",
     "react-bootstrap": "^1.0.1",
     "react-dates": "^21.8.0",
